feat: add Prettier/ESLint dependencies, organize directory, create Hero component, begin CSS process
