+++ src/components/App.js
@@ -1,4 +1,3 @@
-import logo from './logo.svg';
 import './App.css';
 import Hero from './hero/Hero';
 import Footer from './footer/Footer';
