+++ src/components/footer/Footer.js
@@ -1,11 +1,7 @@
 import React from 'react';
-import { Link, NavLink } from 'react-router-dom';
 
 function Footer() {
-
-    return (
-        <section className='footer'>
-            
-        </section>
-    )
+  return <section className="footer" />;
 }
+
+export default Footer;
